project1/Deliverable 3/requirements.txt
# requirements.txt
gradio>=4.0.0
requests
beautifulsoup4
lxml
trafilatura
readability-lxml
transformers
nltk
